Moving the StorageService to the services folder
