+++ src/app/services/location-storage.service.ts
@@ -1,5 +1,5 @@
 import {inject, Injectable} from '@angular/core';
-import {StorageService} from '../generic-cache/storage.service';
+import {StorageService} from './storage.service';
 import {ConditionsAndZip} from '../model/conditions-and-zip.type';
 import {neverExpire} from '../generic-cache/types/no-expiration.type';
